+++ addons/web/static/src/webclient/actions/action_service.js
@@ -86,7 +86,8 @@ export class InvalidButtonParamsError extends Error {}
 // -----------------------------------------------------------------------------
 
 // regex that matches context keys not to forward from an action to another
-const CTX_KEY_REGEX = /^(?:(?:default_|search_default_|show_).+|.+_view_ref|group_by|group_by_no_leaf|active_id|active_ids|orderedBy)$/;
+const CTX_KEY_REGEX =
+    /^(?:(?:default_|search_default_|show_).+|.+_view_ref|group_by|group_by_no_leaf|active_id|active_ids|orderedBy)$/;
 
 // only register this template once for all dynamic classes ControllerComponent
 const ControllerComponentTemplate = xml`<t t-component="Component" t-props="props"/>`;
@@ -839,7 +840,7 @@ function makeActionManager(env) {
             };
             browser.addEventListener("beforeunload", onUnload);
             env.services.ui.block();
-            env.services.router.redirect(action.url);
+            browser.location.assign(action.url);
             browser.removeEventListener("beforeunload", onUnload);
             if (!willUnload) {
                 env.services.ui.unblock();
